+++ src/EscudoDeAtenea.java
@@ -1,5 +1,5 @@
 public class EscudoDeAtenea extends CartaMagica implements Activable{
-    private int boostDef = 800;
+    private short boostDef = 800;
 
     public EscudoDeAtenea() {
         super("Escudo de Atenea", "Aumenta +800 DEF a un monstruo en tu campo.");
