+++ src/LlamadaDelAbismo.java
@@ -1,5 +1,5 @@
 public class LlamadaDelAbismo extends CartaMagica implements Activable {
-    private int costo = 500;
+    private short costo = 500;
 
     public LlamadaDelAbismo() {
         super("Llamada del Abismo",  "Robas 1 carta, pero pierdes 500 LP.");
